Added Forgot Password functionality and a few themes.
These themes are located in theme/variable.scss.  I added in 2 new themes and the original default is still there.  Just uncomment the one you want to be active and comment out the others.  Forgot password feature sendsan email, but the email says "someone requested a password, but it appears you do not have an ionic account".  May need to enable in ionic dashboard or something, not sure.